refactor (LunchItems): 아이템 목록을 다시 render하는 방식을 event를 이용해 상위 요소에게 넘겨주는 방식으로 리팩토링
