+++ src/components/LunchFormItem/LunchFormItem.ts
@@ -1,3 +1,4 @@
+/* eslint-disable prettier/prettier */
 import './style.css';
 
 import { CATEGORIES } from '../../constants/categories';
